==> 1_gerar_db_socio.ipynb <==
import pandas as pd

# LEITURA BASE---------------------------------------------------------------------------------------------------------------------------------------------
df_base = pd.read_csv('/home/matheus/Documentos/relatorio_socios/dados/base_socios.csv', encoding='latin-1', delimiter='##')

# TRATAMENTO PESSOA FÍSICA (NÃO EMPRESA)---------------------------------------------------------------------------------------------------------------------------------
df = df_base.dropna(subset=['DATA_NASCIMENTO'])

# TRATAMENTO SAIDA SOCIEDADE (EMPREENDEDOR ATIVO)-------------------------------------------------------------------------------------------------------------------------------
df = df.set_index(df['DT_SAIDA_SOCIEDADE'])
df['DT_SAIDA_SOCIEDADE'] = df['DT_SAIDA_SOCIEDADE'].astype(str)
df = df.loc[df['DT_SAIDA_SOCIEDADE'].str.contains(r'\b'+'|'.join('NaN') + r'\b')]

# TRATAMENTO UF (MA)----------------------------------------------------------------------------------------------------------------------------------
df = df.set_index(df['COD_UF'])
df = df.loc["MA"]

# TRATAMENTO BASE QUANTIDADE ATRIBUIÇÕES DE EMPREENDEDOR (PESSOA FÍSICA 1/+ ATRIBUIÇÕES)------------------------------------------------------------------------------------------
df_quantidade = pd.DataFrame(data=df)

# TRATAMENTO DE NOMES REPETIDOS (PESSOAS FÍSICAS ÚNICAS)------------------------------------------------------------------------------------------
df_pessoa = pd.DataFrame(data=df)
df_pessoa = df_pessoa.drop_duplicates(subset=['NOME_SOCIO'])


# GERAR NOVO ARQUIVO CSV TRATADO ---------------------------------------------------------------------------------------------------------------------------------
df_pessoa.to_csv('/home/matheus/Documentos/relatorio_socios/dados/df_pessoa/df_pessoa_MA.csv', encoding='latin-1')
df_quantidade.to_csv('/home/matheus/Documentos/relatorio_socios/dados/df_quantidade.csv', encoding='latin-1')
