==> 2_split_nome_socio.ipynb <==
import pandas as pd
df = pd.read_csv('/home/matheus/Documentos/relatorio_socios/dados/df_pessoa/df_pessoa_MA.csv', encoding='latin-1', delimiter=',')

# SPLIT DO NOME_SOCIO/ ADD PRIEIRO NOME NA BD ORIGINAL (df_pessoa_MA.csv)
df_nome = df['NOME_SOCIO'].str.split(" ", n=1, expand=True)
df_nome = df_nome.rename(columns={0:'PRIMEIRO_NOME'})
df.insert(0, 'PRIMEIRO_NOME', df_nome['PRIMEIRO_NOME'])
df
df.to_csv('/home/matheus/Documentos/relatorio_socios/dados/df_pessoa/df_pessoa_nome.csv', encoding='latin-1', sep=',')
