==> 3_nomes_M_F.ipynb <==
import pandas as pd 

df = pd.read_csv('/home/matheus/Documentos/relatorio_socios/dados/dados_brasil_io/grupos.csv', encoding='latin-1', delimiter=',')

# # TRATAMENTO DA BASE DE NOMES MASCUINOS------------------------------------------------------------------------------------------------------
df_M = pd.DataFrame(data=df, columns=('name','classification','names'))
df_M = df_M.set_index(df_M['classification']).loc['M']
df_M = df_M.set_index(df_M['name'])
df_M.to_csv('/home/matheus/Documentos/relatorio_socios/dados/tabela_genero/M/tabela_M.csv',encoding='latin-1', sep=',')
df_M

# TRATAMENTO DA BASE DE NOMES FEMININOS------------------------------------------------------------------------------------------------------
df_F = pd.DataFrame(data=df, columns=('name','classification','names'))
df_F = df_F.set_index(df_F['classification']).loc['F']
df_F = df_F.set_index(df_F['name'])
# df_F.to_csv('/home/matheus/Documentos/relatorio_socios/dados/tabela_genero/F/tabela_F.csv',encoding='latin-1', sep=',')
df_F
