==> 4_sep_M_F.ipynb <==
# ADD LIBS/BASES-----------------------------------------------------------------------------------------------------------------------------
import pandas as pd

df_pessoa = pd.read_csv('/home/matheus/Documentos/relatorio_socios/dados/df_pessoa/df_pessoa_nome.csv', encoding='latin-1', delimiter=',').astype(str)
df_M = pd.read_csv('/home/matheus/Documentos/relatorio_socios/dados/tabela_genero/M/tabela_M.csv', encoding='latin-1', delimiter=',').astype(str)
df_F = pd.read_csv('/home/matheus/Documentos/relatorio_socios/dados/tabela_genero/F/tabela_F.csv', encoding='latin-1', delimiter=',').astype(str)

# REMOÇÃO DE CARACTERES ESPECIAIS ------------------------------------------------------------------------------------------------------------
def correcao_nome(nome):
    nome = (nome.replace('.','').replace('Ç','C').
    replace('Ã','A').replace('Á','A').replace('Â','A').replace('À','A').
    replace('Ẽ','E').replace('É','E').replace('Ê','E').replace('È','E').
    replace('Ĩ','I').replace('Í','I').replace('Î','I').replace('Ì','I').
    replace('Õ','O').replace('Ó','O').replace('Ô','O').replace('Ò','O').
    replace('Ũ','U').replace('Ú','U').replace('Û','U').replace('Ù','U')
    )
    return nome

df_pessoa['PRIMEIRO_NOME'] = df_pessoa['PRIMEIRO_NOME'].astype(str)
df_pessoa['PRIMEIRO_NOME'] = df_pessoa['PRIMEIRO_NOME'].apply(correcao_nome)

# DROP + RENAME --> TABELAS DE NOMES-------------------------------------------------------------------------------------------------------------------------------

# MASCULNO ---------------------------------------------------------------------------------------------------------------------------------------

df_M = df_M.drop(columns={'classification'})
df_M = df_M.rename(columns={'name':'PRIMEIRO_NOME', 'names':'VARIACOES'})
df_M = df_M.drop(columns={'name.1'})
df_M = df_M.astype(str)
# df_M = df_M['PRIMEIRO_NOME'].str.split("|", expand=True)   SPLIT PARA AS VARIAÇÕES DE NOME

# FEMININO ---------------------------------------------------------------------------------------------------------------------------------------

df_F = df_F.drop(columns={'classification'})
df_F = df_F.rename(columns={'name':'PRIMEIRO_NOME', 'names':'VARIACOES'})
df_F = df_F.drop(columns={'name.1'})
df_F = df_F.astype(str)
# df_F = df_F['PRIMEIRO_NOME'].str.split("|", expand=True)   SPLIT PARA AS VARIAÇÕES DE NOME

# VISUALIZAÇÃO --------------------------------------------------------------------------------------------------------------------------------------------------------------

df_F

# MERGE -----------------------------------------------------------------------------------------------------------------------------------------

# MASCULINO ---------------------------------------------------------------------------------------------------------------------------------------
df_teste_M = pd.merge(df_pessoa, df_M['PRIMEIRO_NOME'], how='left', left_on='PRIMEIRO_NOME',right_on='PRIMEIRO_NOME', indicator=True)
df_teste_M = df_teste_M.drop(columns={'Unnamed: 0'})
df_teste_M = df_teste_M.rename(columns={'_merge':'SEXO'})
df_teste_M = df_teste_M.replace('both', 'M')

# # FEMININO ---------------------------------------------------------------------------------------------------------------------------------------
df_teste_F = pd.merge(df_pessoa, df_F['PRIMEIRO_NOME'], how='left', left_on='PRIMEIRO_NOME',right_on='PRIMEIRO_NOME', indicator=True)
df_teste_F = df_teste_F.rename(columns={'_merge':'SEXO'})
df_teste_F = df_teste_F.replace('both', 'F')

# JUNÇÃO M + F BASE FINAL ----------------------------------------------------------------------------------------------------------------------------
df_final = pd.merge(df_teste_M, df_teste_F['SEXO'],  how='left',left_index=True,right_index=True)
df_final = df_final.rename(columns={'SEXO_x':'HOMEM', 'SEXO_y':'MULHER'})
df_final = df_final.drop(columns={'COD_UF.1'})

df_final


# BASE REFERÊNCIA PARA MASCULNO E FEMININO  /////////////  FILTRO PARA SEPARAÇÃO DE NOMES NAO ENCONTRADOS----------------------------------------------------------------------------------------------------------------------------

filtro = (df_final['HOMEM']=='left_only') & (df_final['MULHER'] == 'left_only')
df_nao_definidos = df_final.loc[filtro]
df_final = pd.concat([df_teste_M,df_teste_F], join='inner')


df_final = df_final.set_index(df_final['SEXO'])
df_final = df_final.drop(['left_only'])
df_nao_definidos = df_nao_definidos.drop(columns={'MULHER'})
df_nao_definidos = df_nao_definidos.rename(columns={'HOMEM':'SEXO'})


df_final

# GERADOR CSV ----------------------------------------------------------------------------------------------------------------------------
df_final.to_csv('/home/matheus/Documentos/relatorio_socios/dados/Base_Final.csv', encoding='latin-1', sep=',')
